requirements.txt
streamlit
ultralytics
opencv-python-headless
torch
torchvision
Pillow
numpy
